+++ src/main/scala/edu/ie3/simona/agent/grid/powerflow/DBFSAlgorithm.scala
@@ -14,7 +14,11 @@ import edu.ie3.powerflow.model.PowerFlowResult
 import edu.ie3.powerflow.model.PowerFlowResult.FailedPowerFlowResult.FailedNewtonRaphsonPFResult
 import edu.ie3.powerflow.model.PowerFlowResult.SuccessFullPowerFlowResult.ValidNewtonRaphsonPFResult
 import edu.ie3.powerflow.model.enums.NodeType
-import edu.ie3.simona.agent.grid.GridAgent.{afterPowerFlow, unsupported}
+import edu.ie3.simona.agent.grid.GridAgent.{
+  afterPowerFlow,
+  askInferior,
+  unsupported,
+}
 import edu.ie3.simona.agent.grid.GridAgentMessages.*
 import edu.ie3.simona.agent.grid.GridAgentMessages.Responses.{
   ExchangePower,
@@ -95,9 +99,14 @@ trait DBFSAlgorithm extends PowerFlowSupport with GridResultsSupport {
             gridAgentBaseData.gridEnv.gridModel.mainRefSystem,
           )
           // 2. inferior grids p/q values
-          askInferiorGridsForPowers(
-            gridAgentBaseData.currentSweepNo,
+          askInferior(
             gridAgentBaseData.inferiorGridRefs,
+            (ref, inferiorGridNodes) =>
+              RequestGridPower(
+                gridAgentBaseData.currentSweepNo,
+                inferiorGridNodes,
+                ref,
+              ),
           )
 
           // 3. superior grids slack voltage
@@ -652,9 +661,14 @@ trait DBFSAlgorithm extends PowerFlowSupport with GridResultsSupport {
 
                 // 2. inferior grids p/q values
                 val askForInferiorGridPowersOpt =
-                  askInferiorGridsForPowers(
-                    updatedGridAgentBaseData.currentSweepNo,
-                    updatedGridAgentBaseData.inferiorGridRefs,
+                  askInferior(
+                    gridAgentBaseData.inferiorGridRefs,
+                    (ref, inferiorGridNodes) =>
+                      RequestGridPower(
+                        updatedGridAgentBaseData.currentSweepNo,
+                        inferiorGridNodes,
+                        ref,
+                      ),
                   )
 
                 // when we don't have inferior grids and no assets both methods return None, and we can skip doing another power
@@ -1119,36 +1133,6 @@ trait DBFSAlgorithm extends PowerFlowSupport with GridResultsSupport {
     case (ctx, msg) =>
       unsupported(msg, ctx.log)
       Behaviors.same
-  }
-
-  /** Triggers an execution of the pekko `ask` pattern for all power values @
-    * connection nodes of inferior grids (if any) of this [[GridAgent]].
-    *
-    * @param currentSweepNo
-    *   The current sweep number the DBFS is in.
-    * @param inferiorGridRefs
-    *   A map containing a mapping from [[ActorRef]]s to corresponding [[UUID]]s
-    *   of inferior nodes.
-    * @return
-    *   True if this grids has connected inferior grids or false if this no
-    *   inferior grids.
-    */
-  private def askInferiorGridsForPowers(
-      currentSweepNo: Int,
-      inferiorGridRefs: Map[ActorRef[GridAgent.Message], Set[UUID]],
-  )(using ctx: ActorContext[GridAgent.Message]): Boolean = {
-    if inferiorGridRefs.nonEmpty then {
-      inferiorGridRefs.foreach {
-        case (inferiorGridAgentRef, inferiorGridGateNodes) =>
-          inferiorGridAgentRef ! RequestGridPower(
-            currentSweepNo,
-            inferiorGridGateNodes,
-            ctx.self,
-          )
-      }
-
-      true
-    } else false
   }
 
   /** Triggers an execution of the pekko `ask` pattern for all slack voltages of